% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birth_death_stochastic.R
\name{birth_death_stochastic}
\alias{birth_death_stochastic}
\title{Euler-Maruyama method birth-death solution for a stochastic differential equation.}
\usage{
birth_death_stochastic(
  birth_rate,
  death_rate,
  init_cond,
  parameters = NULL,
  t_start = 0,
  deltaT = 1,
  n_steps = 1,
  sigma = 1
)
}
\arguments{
\item{birth_rate}{The 1 or multi dimensional system of equations for the birth rate, written in formula notation as a vector (i.e.  c(dx ~ f(x,y), dy ~ g(x,y)))}

\item{death_rate}{The 1 or multi dimensional system of equations for the death rate of the differential equation, written in formula notation as a vector (i.e.  c(dx ~ f(x,y), dy ~ g(x,y)))}

\item{init_cond}{Listing of initial conditions (we can only do one initial condition)}

\item{parameters}{The values of the parameters we are using}

\item{t_start}{The starting time point (defaults to t = 0)}

\item{deltaT}{The timestep length (defaults to 1)}

\item{n_steps}{The number of timesteps to compute solution (defaults to n_steps = 1)}

\item{sigma}{scaling factor for of the stochastic part of the SDE}
}
\value{
A tidy of data frame the solutions
}
\description{
\code{birth_death_stochastic} solves a multi-dimensional differential equation using a birth-death process, applying the Euler-Maruyama method. A reality constraint is applied so the variables can never be zero.
}
\examples{
TBD
}
