% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthDeathStochasticSystems.R
\name{birthDeathStochasticSystems}
\alias{birthDeathStochasticSystems}
\title{Euler Maruyama method solution for a stochastic system of differential equations.}
\usage{
birthDeathStochasticSystems(deltaT = 1, timeSteps = 1,
  initialCondition, FUN = birth, FUN_2 = death,
  parameters = parameters, nSimulations = 1, sigma = 1)
}
\arguments{
\item{deltaT}{Size of timesteps}

\item{timeSteps}{Number of timesteps we solve.  deltaT*timeSteps = total time}

\item{initialCondition}{Listing of initial conditions (we can only do one initial condition)}

\item{parameters}{The values of the parameters we are using}

\item{nSimulations}{the number of simulations we are running}

\item{sigma}{scaling factor for of the stochastic part of the SDE}

\item{birth}{a function that we have for the ``birth'' processes of the SDE}

\item{death}{a function we have for the ``death'' processes of the SDE}
}
\value{
A spaghetti plot and ensemble average plot of your solution
}
\description{
\code{birthDeathStochasticSystems} solves a multi-dimensional differential equation with the Euler Maruyama method with stochastic elements using a birth death process.
}
