% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_model.R
\name{bootstrap_model}
\alias{bootstrap_model}
\title{\code{bootstrap_model} computes a given number of bootstrap samples, computes a linear model, and then plots.  The resulting histogram of samples is shown, along with a sampling of the results}
\source{
Code is adapted from: \url{https://cran.r-project.org/web/packages/broom/vignettes/bootstrapping.html}
}
\usage{
bootstrap_model(data, regression_formula, n = 100, x_label = "x",
  y_label = "y")
}
\arguments{
\item{data}{two column of data to be plotted.
First column is independent variable. Second column dependent variable.}

\item{regression_formula}{function you use to fit the data}

\item{n}{number of bootstrap samples}

\item{x_label}{value for the x label axis}

\item{y_label}{value for the y label axis}
}
\description{
\code{bootstrap_model} computes a given number of bootstrap samples, computes a linear model, and then plots.  The resulting histogram of samples is shown, along with a sampling of the results
}
\examples{

# Identify the name of the data set you wish to run a regression on and labels
for the axis.
my_data <- data.frame(age=Loblolly$age,height=Loblolly$height)

#' Identify the basic formula of the regression and plotting formula. Make sure you have the same names of the columns in your data.
regression_formula = height ~ 1 + age

bootstrap_model(my_data,regression_formula,n=100,'Age','Height')
}
