% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_samples.R
\name{bootstrap_samples}
\alias{bootstrap_samples}
\title{\code{bootstrap_samples} computes a given number of bootstrap samples and computes the distribution of bootstrap results.  The resulting histogram of samples is shown, along with a sampling of the results}
\source{
Code is adapted from: \url{https://cran.r-project.org/web/packages/broom/vignettes/bootstrapping.html}
}
\usage{
bootstrap_samples(data, n = 100, x_label = "x")
}
\arguments{
\item{data}{one column of data to take bootstrap samples of.}

\item{n}{number of bootstrap samples}

\item{x_label}{value for the x label axis}
}
\description{
\code{bootstrap_samples} computes a given number of bootstrap samples and computes the distribution of bootstrap results.  The resulting histogram of samples is shown, along with a sampling of the results
}
\examples{

# Identify the name of the data set you wish to run a regression on and labels
for the axis.
my_data <- data.frame(precipitation$precip)

#' Identify the basic formula of the regression and plotting formula. Make sure you have the same names of the columns in your data.
regression_formula = height ~ 1 + age

bootstrap_samples(my_data,n=100,'Precipitation')
}
