% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casino.R
\name{casino}
\alias{casino}
\title{Simulate a casino game}
\usage{
casino(houseWinProbability, initialCash, nTimes, nSimulations)
}
\arguments{
\item{houseWinProbability}{probability the house wins the game}

\item{initialCash}{cash you start out with}

\item{nTimes}{times you decide to play the game}

\item{nSimulations}{the number of times you want to simulate the results}
}
\value{
A spaghetti plot and ensemble average plot of your winnings
}
\description{
\code{casino} visualizes the vector field for a one or two dimensional differential equation.
}
\examples{

casino(.52,100,200,100)
}
