% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggquiver.R
\docType{package}
\name{demodelr-package}
\alias{demodelr}
\alias{demodelr-package}
\title{demodelr: Simulating differential equations with data.}
\description{
A collection of files for use with the text Exploring modeling with data and differential equations using R.
}
\keyword{internal}
