% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenvalues.R
\name{eigenvalues}
\alias{eigenvalues}
\title{Matrix eigenvalues and eigenvectors}
\usage{
eigenvalues(matrix_entries, matrix_rows = 2)
}
\arguments{
\item{matrix_entries}{entries of your matrix in row wise format.
So the matrix
#   4  3
#   2  1
# would be entered in c(4,3,2,1)}

\item{matrix_rows}{the number of rows and columns in your SQUARE matrix.}
}
\value{
The result is a list with two elements (denoted by the “$”), values and vectors.
result$values are the eigenvalues, stored as a vector. The leading eigenvalue is the first entry in the vector.
}
\description{
\code{eigenvalues} visualizes the vector field for a one or two dimensional differential equation.
}
\examples{
eigenvalues(c(1,2,3,4))

# Note: for the 3 x 3 case, we need to define the number of matrix rows:
eigenvalues(c(1,2,3,4,5,6,7,8,9),matrix_rows=3)
}
