% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euler.R
\name{euler}
\alias{euler}
\title{Euler's method solution for a differential equation.}
\usage{
euler(
  system_eq,
  initial_condition,
  parameters = NULL,
  t_start = 0,
  deltaT = 1,
  n_steps = 1
)
}
\arguments{
\item{system_eq}{(REQUIRED) The 1 or multi dimensional system of equations, written in formula notation as a vector (i.e.  c(dx ~ f(x,y), dy ~ g(x,y)))}

\item{initial_condition}{(REQUIRED) Listing of initial conditions, as a vector}

\item{parameters}{The values of the parameters we are using (optional)}

\item{t_start}{The starting time point (defaults to t = 0)}

\item{deltaT}{The timestep length (defaults to 1)}

\item{n_steps}{The number of timesteps to compute solution (defaults to n_steps = 1)}

\item{timeSteps}{Number of timesteps we solve.  deltaT*timeSteps = total time}
}
\value{
A tidy of data frame the solutions
}
\description{
\code{euler} solves a multi-dimensional differential equation with Euler's method.  The parameters listed as required are needed
See the vignette for detailed examples of usage.
}
\examples{
# Run the vignette that works through an example
vignette("eulers-method")
}
