% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euler_stochastic.R
\name{euler_stochastic}
\alias{euler_stochastic}
\title{Euler-Maruyama method solution for a stochastic differential equation.}
\usage{
euler_stochastic(
  deterministic_rate,
  stochastic_rate,
  init_cond,
  parameters = NULL,
  t_start = 0,
  deltaT = 1,
  n_steps = 1,
  D = 1
)
}
\arguments{
\item{deterministic_rate}{The 1 or multi dimensional system of equations for the deterministic part of the differential equation, written in formula notation as a vector (i.e.  c(dx ~ f(x,y), dy ~ g(x,y)))}

\item{stochastic_rate}{The 1 or multi dimensional system of equations for the stochastic part of the differential equation, written in formula notation as a vector (i.e.  c(dx ~ f(x,y), dy ~ g(x,y)))}

\item{init_cond}{Listing of initial conditions (we can only do one initial condition)}

\item{parameters}{The values of the parameters we are using}

\item{t_start}{The starting time point (defaults to t = 0)}

\item{deltaT}{The timestep length (defaults to 1)}

\item{n_steps}{The number of timesteps to compute solution (defaults to n_steps = 1)}

\item{D}{diffusion coefficient for the stochastic part of the SDE}
}
\value{
A tidy of data frame the solutions
}
\description{
\code{euler_stochastic} solves a multi-dimensional differential equation with the Euler-Maruyama method with stochastic elements.
}
\examples{
TBD
}
