% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eulerStochastic.R
\name{eulerStochastic}
\alias{eulerStochastic}
\title{Euler's method solution for a differential equation.}
\usage{
eulerStochastic(deltaT = 1, timeSteps = 1, initialCondition,
  FUN = deterministicDynamics, FUN_2 = stochasticDynamics,
  parameters = parameters, nSimulations = 1, sigma = 1)
}
\arguments{
\item{deltaT}{Size of timesteps}

\item{timeSteps}{Number of timesteps we solve.  deltaT*timeSteps = total time}

\item{initialCondition}{Listing of initial conditions (we can only do one initial condition)}

\item{parameters}{The values of the parameters we are using}

\item{nSimulations}{the number of simulations we are running}

\item{sigma}{scaling factor for of the stochastic part of the SDE}

\item{deterministicDynamics}{a function that we have for the deterministic part of the SDE}

\item{stochasticDynamics}{a function we have for the stochastic part of the SDE}
}
\value{
A spaghetti plot and ensemble average plot of your solution
}
\description{
\code{eulerStochastic} solves a multi-dimensional differential equation with Euler's method with stochastic elements.
See the vignette for detailed examples of usage.
}
\examples{
# Run the vignette that works through an example:
vignette("euler-stochastic")
}
