% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_temperature.R
\docType{data}
\name{global_temperature}
\alias{global_temperature}
\title{Measured average global temperature by year}
\format{
A data frame with 135 rows and 2 variables
}
\source{
The data were collected from NOAA. \url{https://climate.nasa.gov/vital-signs/global-temperature/}
}
\usage{
data(global_temperature)
}
\description{
A dataset containing phosophorous content in Daphnia and algae. The variables are as follows:
}
\details{
\itemize{
  \item yearSince1880. The year since 1880 (year)
  \item globalTemp. Average global temperature (degrees Celsius)
}
}
\keyword{datasets}
