% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informationCriteria.R
\name{informationCriteria}
\alias{informationCriteria}
\title{\code{informationCriteria} fits a functional relationship to data, summarizes the results, makes a plot of the fitted formula, and reports values needed to compute the AIC and the BIC.}
\usage{
informationCriteria(data, regression_formula, x_label = "x", y_label = "y")
}
\arguments{
\item{data}{two column of data to be plotted.
First column is independent variable. Second column dependent variable.}

\item{regression_formula}{function you use to fit the data}

\item{x_label}{value for the x label axis}

\item{y_label}{value for the y label axis}
}
\description{
\code{informationCriteria} fits a functional relationship to data, summarizes the results, makes a plot of the fitted formula, and reports values needed to compute the AIC and the BIC.
}
\examples{
# Identify the name of the data set you wish to run a regression on and labels
for the axis.
my_data <- data.frame(age=Loblolly$age,height=Loblolly$height)

#' Identify the basic formula of the regression and plotting formula. Make sure you have the same names of the columns in your data.
regression_formula = height ~ 1 + age+ I(age^2)

informationCriteria(my_data,regression_formula,'Age','Height')
}
