% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggquiver.R
\docType{package}
\name{MAT369Code-package}
\alias{MAT369Code}
\alias{MAT369Code-package}
\title{MAT369Code: Simulating differential equations with data.}
\description{
A collection of files for use in modeling with data and differential equations.
    Will write more here.
}
\keyword{internal}
