% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_analyze.R
\name{mcmc_analyze}
\alias{mcmc_analyze}
\title{Markov Chain parameter estimates}
\usage{
mcmc_analyze(
  model,
  data,
  mcmc_out,
  mode = "emp",
  initial_condition = NULL,
  deltaT = NULL,
  n_steps = NULL
)
}
\arguments{
\item{model}{the model equations that we use to compute the result.}

\item{data}{the data used to assess the model}

\item{mcmc_out}{A dataframe: the first column is the accept flag  of the mcmc run (TRUE/FALSE), the log likelihood, and the parameter values}

\item{mode}{two choices: emp --> empirical (default) or de --> differential equations. The estimator works differently depending on which is used.}

\item{initial_condition}{The initial condition for the differential equation (DE mode only)}

\item{deltaT}{The length between timesteps (DE mode only)}

\item{n_steps}{The number of time steps we run the model (DE mode only)}
}
\description{
\code{mcmc_analyze} Computes summary histograms and model-data comparisons from and Markov Chain Monte Carlo parameter estimate for a given model
}
