% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_estimate.R
\name{mcmc_estimate}
\alias{mcmc_estimate}
\title{Markov Chain parameter estimates}
\usage{
mcmc_estimate(
  model,
  data,
  parameters,
  iterations = 1,
  knob_flag = FALSE,
  mode = "emp",
  initial_condition = NULL,
  deltaT = NULL,
  n_steps = NULL
)
}
\arguments{
\item{model}{the model equations that we use to compute the result.}

\item{data}{the data used to assess the model}

\item{parameters}{a data frame that lists the names of the parameters along with upper and lower bounds}

\item{iterations}{the number of iterations we wish to run the MCMC for.}

\item{knob_flag}{determines if we tune the range that can be search (annealling)}

\item{mode}{two choices: emp --> empirical (default) or de --> differential equations. The estimator works differently depending on which is used.}

\item{initial_condition}{The initial condition for the differential equation (DE mode only)}

\item{deltaT}{The length between timesteps (DE mode only)}

\item{n_steps}{The number of time steps we run the model (DE mode only)}
}
\value{
A dataframe: the first column is the accept flag  of the mcmc run (TRUE/FALSE), the log likelihood, and the parameter values
}
\description{
\code{mcmc_estimate} Computes and Markov Chain Monte Carlo parameter estimate for a given model
}
