% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modCost_JZ.R
\name{modCost_JZ}
\alias{modCost_JZ}
\title{Markov Chain parameter estimates}
\usage{
modCost_JZ(model, obs, x = "time", y = NULL, err = NULL,
  weight = "none", scaleVar = FALSE, cost = NULL, ...)
}
\description{
\code{modCost_JZ} Computes and Markov Chain Monte Carlo parameter estimate for a given model.  This is a change from the code provided in the FME package to account for the likelihood.
}
