% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parks.R
\docType{data}
\name{parks}
\alias{parks}
\title{Visitor and resource usage to a national park}
\format{
A data frame with 8 rows and 3 variables
}
\source{
Laura Sinay, Leon Sinay. 2006. “A Simple Mathematical Model for the Effects of the Growth of Tourism on Environment.”
}
\usage{
data(parks)
}
\description{
A dataset containing phosophorous content in Daphnia and algae. The variables are as follows:
}
\details{
\itemize{
  \item time. (days)
  \item vistors. number of visitors to a national parked, scaled by the equilibrium value.
  \item resources. scaled area of the reserve not deforested.
}
}
\keyword{datasets}
