% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phaseplane.R
\name{phaseplane}
\alias{phaseplane}
\title{Phase plane of differential equation.}
\usage{
phaseplane(n_points, x_window, y_window, x_label, y_label, dx, dy)
}
\arguments{
\item{n_points}{number of points we evaluate on the grid in both directions}

\item{x_window}{x axis limits.  Must be of the form c(minVal,maxVal)}

\item{y_window}{y axis limits.  Must be of the form c(minVal,maxVal)}

\item{x_label}{x axis label for plot}

\item{y_label}{y axis label for plot}

\item{dx}{a function for the dx/dt = f(x,y)}

\item{dy}{a function for the dy/dt = f(x,y)}

\item{initialCondition}{Listing of initial conditions}
}
\value{
A phase plane diagram of system of differential equations
}
\description{
\code{phaseplane} visualizes the vector field for a one or two dimensional differential equation.
}
\examples{
# For a two variable system of differential equations we define the dx/dt and the dy/dt separately:
dx <- function(x,y) {
cos(x)
}

dy <- function(x,y) {
 sin(y)
}

phaseplane(20,c(-3,3),c(-3,3),'X','Y',dx,dy)

# For a one dimensional system: dy/dt = f(t,y).  In this case the xWindow represents time.
# However, we still need to write dx as a function of x and y, but the plot can be labeled T

dy <- function(x,y) {
-y
}

phaseplane(20,c(-3,3),c(-3,3),'T','Y',1,dy)

# Here is a second example for the differential equation dy/dt = -cos(t) y
# In this case, the x coordinate represents time.

dy <- function(x,y) {
-cos(x)*y
}

phaseplane(20,c(-3,3),c(-3,3),'T','Y',1,dy)

}
