% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotData.R
\name{plotData}
\alias{plotData}
\title{Plot a simple algebraic equation with data}
\usage{
plotData(data, x_label = "x", y_label = "y")
}
\arguments{
\item{data}{two column of data to be plotted.
First column is independent variable. Second column dependent variable.
Must be a data.frame}

\item{x_label}{value for the x label axis}

\item{y_label}{value for the y label axis}
}
\value{
A plot of your function
}
\description{
\code{plotData} plots an x-y relationship with data
}
\examples{
data <- data.frame(Loblolly$age,Loblolly$height)
plotData(data,'age','height')
}
