% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunction.R
\name{plotFunction}
\alias{plotFunction}
\title{Plot a simple algebraic equation}
\usage{
plotFunction(x, y, x_label = "x", y_label = "y")
}
\arguments{
\item{x}{The independent variable}

\item{y}{The dependent variable}

\item{x_label}{value for the x label axis}

\item{y_label}{value for the y label axis}
}
\value{
A plot of your function
}
\description{
\code{plotFunction} plots an x-y relationship of data
See the vignette for detailed examples of usage.
}
\examples{
x<- 1:10
y<- x^2
plotFunction(x,y,'x','y')
}
