% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunction_Data.R
\name{plotFunction_Data}
\alias{plotFunction_Data}
\title{Plot a simple algebraic equation with data}
\usage{
plotFunction_Data(x, y, data, x_label = "x", y_label = "y")
}
\arguments{
\item{x}{The independent variable}

\item{y}{The dependent variable}

\item{data}{two column of data to be plotted.
First column is independent variable. Second column dependent variable.
Must be a data.frame}

\item{x_label}{value for the x label axis}

\item{y_label}{value for the y label axis}
}
\value{
A plot of your function
}
\description{
\code{plotFunction_Data} plots an x-y relationship with data
}
\examples{
x<- 0:25
y<- x^2
data <- data.frame(Loblolly$age,Loblolly$height)
plotFunction_Data(x,y,data,'x','y')
}
