% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHistogram.R
\name{plotHistogram}
\alias{plotHistogram}
\title{Plot a histogram of a dataset}
\usage{
plotHistogram(data, bins = 10, x_label = "x")
}
\arguments{
\item{data}{two column of data to be plotted.
First column is independent variable. Second column dependent variable.
Must be a data.frame}

\item{bins}{number of bins you wish for the data}

\item{x_label}{value for the x label axis}
}
\value{
A plot of your function
}
\description{
\code{plotHistogram} plots a histogram of a dataset
}
\examples{
data <- data.frame(Loblolly$height)
plotHistogram(data,10,'height')
}
