% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLikelihood.R
\name{plotLikelihood}
\alias{plotLikelihood}
\title{Likelihood plot of a two parameter model}
\usage{
plotLikelihood(model, data, parameters, logLikely = FALSE)
}
\arguments{
\item{model}{a function or model of our situation}

\item{data}{two column of data to be plotted.
First column is independent variable. Second column dependent variable.
Must be a data.frame}

\item{parameters}{The values of the parameters we are using}

\item{logLikely}{Do we compute the log likelihood function (default is FALSE)}
}
\value{
A plot of your likelihood function, with optimum value identified
}
\description{
\code{plotLikelihood} makes a likelihood plot of a two parameter model.
}
\examples{
# Run the vignette that works through an example
vignette("likelihood")

}
