% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRegression_Data.R
\name{plotRegression_Data}
\alias{plotRegression_Data}
\title{\code{plotRegression_Data} fits a functional relationship to data, summarizes the results, and makes a plot of the fitted formula.}
\usage{
plotRegression_Data(data, regression_formula, x_label = "x",
  y_label = "y")
}
\arguments{
\item{data}{two column of data to be plotted.
First column is independent variable. Second column dependent variable.
Must be a data.frame, labeled as x and y.}

\item{regression_formula}{function you use to fit the data}

\item{x_label}{value for the x label axis}

\item{y_label}{value for the y label axis}
}
\description{
\code{plotRegression_Data} fits a functional relationship to data, summarizes the results, and makes a plot of the fitted formula.
}
\examples{
# Identify the name of the data set you wish to run a regression on and labels
for the axis.  You must label the columns as x and y.

#' Identify the basic formula of the regression and plotting formula. Make sure you have the same names of the columns in your data.
regression_formula = height ~ 1 + age+ I(age^2)

plotRegression_Data(Loblolly,regression_formula,'Height','Age')
}
