% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{population}
\alias{population}
\title{Simulate a population dynamics that produce offspring}
\usage{
population(initialPopulation, nGenerations, nSimulations)
}
\arguments{
\item{initialPopulation}{Initial population size}

\item{nGenerations}{the number of Generations}

\item{nSimulations}{Number of simulations we try}
}
\value{
A spaghetti plot and ensemble average plot of population dynamics
}
\description{
\code{population} modified from Logan and Wolesesnky "Mathematical Methods in Biology". Simulates the dynamics of population that produces variable number of offspring generation to generation. This corresponds to Exercise 5 on page 318
}
\examples{

population(10,200,10)
}
