% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precipitation.R
\docType{data}
\name{precipitation}
\alias{precipitation}
\title{Measured precipitation from a rainfall event}
\format{
A data frame with 56 rows and 5 variables
}
\source{
The data were collected from Community Collaborative Rain Hail and Snow Network (CoCoRaHS). \url{https://www.cocorahs.org/ViewData/ListDailyPrecipReports.aspx}
}
\usage{
data(precipitation)
}
\description{
A dataset containing measured precipitation data from the Minneapolis St. Paul Area:
}
\details{
\itemize{
  \item date. Calendar day of year of measurement
  \item time. Time measurement is made
  \item station_id Shorthand name for station in CoCoRaHS network
  \item station_name Name of station in CoCoRaHS network
  \item precip. Observed precipitation (inches)
}
}
\keyword{datasets}
