% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomWalk.R
\name{randomWalk}
\alias{randomWalk}
\title{Simulate a one dimensional random walk}
\usage{
randomWalk(steps, nSimulations)
}
\arguments{
\item{steps}{number of steps you run the walk}

\item{nSimulations}{the number of times you want to simulate the results}
}
\value{
A spaghetti plot and ensemble average plot of your winnings
}
\description{
\code{randomWalk} does a 1-dimensional random walk where you are equally likely to move to the left or the right.
}
\examples{

randomWalk(100,200)
}
