% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rk4.R
\name{rk4}
\alias{rk4}
\title{Runge Kutta method solution for a differential equation.}
\usage{
rk4(
  system_eq,
  initial_condition,
  parameters = NULL,
  t_start = 0,
  deltaT = 1,
  n_steps = 1
)
}
\arguments{
\item{system_eq}{(REQUIRED) The 1 or 2 dimensional system of equations, written in formula notation as a vector (i.e.  c(dx ~ f(x,y), dy ~ g(x,y)))}

\item{initial_condition}{(REQUIRED) Listing of initial conditions, as a vector}

\item{parameters}{The values of the parameters we are using (optional)}

\item{t_start}{The starting time point (defaults to t = 0)}

\item{deltaT}{The timestep length (defaults to 1)}

\item{n_steps}{The number of timesteps to compute solution (defaults to n_steps = 1)}

\item{timeSteps}{Number of timesteps we solve.  deltaT*timeSteps = total time}
}
\value{
A tidy of data frame the solutions
}
\description{
\code{rk4} solves a multi-dimensional differential equation with Runge-Kutta 4th order method.  The parameters listed as required are needed
See the vignette for detailed examples of usage.
}
\references{
See https://en.wikipedia.org/wiki/Runge%E2%80%93Kutta_methods
}
