% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandhillsDemographic.R
\name{sandhillsDemographic}
\alias{sandhillsDemographic}
\title{Simulate population dynamics for sandhills cranes}
\usage{
sandhillsDemographic(initialPopulation, nYears, nSimulations)
}
\arguments{
\item{initialPopulation}{Initial population size}

\item{nYears}{the number of years we run our population}

\item{nSimulations}{Number of simulations we try}
}
\value{
A spaghetti plot and ensemble average plot of population dynamics
}
\description{
\code{sandhillsDemographic} is modified from Logan and Wolesesnky "Mathematical Methods in Biology." Create a sandhill crane model of discrete dynamics x(t+1)=r*x(t), as detailed on page 311.  The parameter r = 1+b-d, where b is the birth rate, d is the death rate. b and d are drawn from normally distributed random variables.
}
\examples{

sandhillsDemographic(100,20,10)
}
