% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandhillsEnvironmental.R
\name{sandhillsEnvironmental}
\alias{sandhillsEnvironmental}
\title{Simulate population dynamics for sandhills cranes}
\usage{
sandhillsEnvironmental(initialPopulation, floodRate, nYears, nSimulations)
}
\arguments{
\item{initialPopulation}{Initial population size}

\item{floodRate}{The frequency of flooding. (1/25 = one in 25 years)}

\item{nYears}{the number of years we run our population}

\item{nSimulations}{Number of simulations we try}
}
\value{
A spaghetti plot and ensemble average plot of population dynamics
}
\description{
\code{sandhillsEnvironmental} is modified from Logan and Wolesesnky "Mathematical Methods in Biology." Create a sandhill crane model of discrete dynamics x(t+1)=r*x(t), as detailed on page 311.  The parameter r = 1+b-d, where b is the birth rate, d is the death rate. We modify r based on the stated flood rate potential. In normal years, the value of r is 1.4. In Catastrophic years the net growth rate is lowered so r=0.575.
}
\examples{

sandhillsEnvironmental(100,1/25,20,10)
}
