% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowfall.R
\docType{data}
\name{snowfall}
\alias{snowfall}
\title{Measured snowfall from a blizzard in April 2018}
\format{
A data frame with 16 rows and 5 variables
}
\source{
The data were collected from Community Collaborative Rain Hail and Snow Network (CoCoRaHS). \url{https://www.cocorahs.org/ViewData/ListDailyPrecipReports.aspx}
}
\usage{
data(snowfall)
}
\description{
A dataset containing measured snowfall data from the Minneapolis St. Paul Area:
}
\details{
\itemize{
  \item date. Calendar day of year of measurement
  \item time. Time measurement is made
  \item station_id Shorthand name for station in CoCoRaHS network
  \item station_name Name of station in CoCoRaHS network
  \item snowfall total snowfall (inches)
}
}
\keyword{datasets}
