% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeDynamics.R
\name{summarizeDynamics}
\alias{summarizeDynamics}
\title{Simulate a one dimensional random walk}
\usage{
summarizeDynamics(simulations, labels)
}
\arguments{
\item{simulations}{data frame from a stochastic simulation}

\item{labels}{the labes on the plot axes}
}
\value{
A spaghetti plot and ensemble average plot of simulation
}
\description{
\code{summarizeDynamics} takes the result from a stochastic simulation and produces a spaghetti plot and ensemble average plot.
}
\examples{

# Run the vignette that works through an example
vignette("stochstic-logistic")
}
