% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systems.R
\name{systems}
\alias{systems}
\title{Numerical solution for a differential equation.}
\usage{
systems(deltaT = 1, timeSteps = 1, initialCondition, FUN = dynamics,
  parameters = parameters, returnData = FALSE)
}
\arguments{
\item{deltaT}{Size of timesteps}

\item{timeSteps}{Number of timesteps we solve.  deltaT*timeSteps = total time}

\item{initialCondition}{Listing of initial conditions}

\item{parameters}{The values of the parameters we are using}

\item{returnData}{A T/F if we want to return just the solution data than a plot.}

\item{dynamics}{a Function that we have for our dynamics}
}
\value{
A plot of your solved equation solution
}
\description{
\code{systems} solves a multi-dimensional dimensional differential equation with numerical differential equation solvers.
See the vignette for detailed examples of usage.
}
\examples{
systems(deltaT,timeSteps,initialCondition,FUN=dynamics,parameters=parameters)
# Run the vignette that works through an example
vignette("systems")
}
