% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildebeast.R
\name{wildebeast}
\alias{wildebeast}
\title{Simulate wildebeast population dynamics}
\usage{
wildebeast(harvestRate, nYears, nSimulations)
}
\arguments{
\item{harvestRate}{The harvesting rate of the wildebeast.}

\item{nYears}{Number of years we run the population}

\item{nSimulations}{Number of simulations we try}
}
\value{
A spaghetti plot and ensemble average plot of population dynamics
}
\description{
\code{wildebeast} is modified from Logan and Wolesesnky "Mathematical Methods in Biology". Simluate a wildebeast population undergoing harvesting with stochastic environmental effects. due to rainfall. This corresponds to Exercise 8 on page 318
}
\examples{

wildebeast(.5,5,2)
}
