% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilson.R
\docType{data}
\name{wilson}
\alias{wilson}
\title{Measured Weight of a dog over time}
\format{
A data frame with 19 rows and 2 variables
}
\source{
From \url{https://bscheng.com/2014/05/07/modeling-logistic-growth-data-in-r/}
}
\usage{
data(wilson)
}
\description{
A dataset containing the mass of a growing dog.
}
\details{
\itemize{
  \item days since birth
  \item weight. (pounds)
}
}
\keyword{datasets}
