% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yeast.R
\docType{data}
\name{yeast}
\alias{yeast}
\title{Measured Sacchromyces data (yeast) from Gause 1932 "Experimental studies on the struggle for coexistence"}
\format{
A data frame with 7 rows and 2 variables
}
\source{
Table 1 from Gause, G. F. 1932. “Experimental Studies on the Struggle for Existence: I. Mixed Population of Two Species of Yeast.” Journal of Experimental Biology 9 (4): 389–402.
}
\usage{
data(yeast)
}
\description{
A dataset containing measurements of growth of yeast in a culture. The variables are as follows:
}
\details{
\itemize{
  \item time. (hours)
  \item volume. Sacchromyces volume in container (cubic centimeters)
}
}
\keyword{datasets}
