% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euler.R
\name{euler}
\alias{euler}
\title{Solve a multi-dimensional dimensional differential equation with Euler's method. See the vignette for detailed examples of usage.}
\usage{
euler(deltaT, timeSteps, variableNames, FUN = dynamics,
  parameters = parameters)
}
\arguments{
\item{deltaT}{Size of timesteps}

\item{timeSteps}{Number of timesteps we solve.  deltaT*timeSteps = total time}

\item{variableNames}{Name of variables (i.e. P,V) that we have in our differential equation.}

\item{parameters}{The values of the parameters we are using}

\item{dynamics}{a Function that we have for our dynamics}
}
\value{
A plot of your Euler's method solution
}
\description{
Solve a multi-dimensional dimensional differential equation with Euler's method. See the vignette for detailed examples of usage.
}
\examples{
euler(deltaT,timeSteps,variableNames,FUN=dynamics,parameters=parameters)
}
