% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euler.R
\name{euler}
\alias{euler}
\title{Euler's method solution for a differential equation.}
\usage{
euler(
  deltaT = 1,
  timeSteps = 1,
  initialCondition,
  FUN = dynamics,
  parameters = parameters,
  returnData = FALSE
)
}
\arguments{
\item{deltaT}{Size of timesteps}

\item{timeSteps}{Number of timesteps we solve.  deltaT*timeSteps = total time}

\item{initialCondition}{Listing of initial conditions}

\item{parameters}{The values of the parameters we are using}

\item{returnData}{A T/F if we want to return just the solution data than a plot.}

\item{dynamics}{a Function that we have for our dynamics}
}
\value{
A plot of your Euler's method solution
}
\description{
\code{euler} solves a multi-dimensional differential equation with Euler's method.
See the vignette for detailed examples of usage.
}
\examples{
euler(deltaT,timeSteps,initialCondition,variableNames,FUN=dynamics,parameters=parameters)
# Run the vignette that works through an example
vignette("eulers-method")
}
