% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euler.R
\name{euler}
\alias{euler}
\title{Euler's method solution for a differential equation.}
\usage{
euler(deltaT = 1, timeSteps = 1, initialCondition, variableNames,
  FUN = dynamics, parameters = parameters)
}
\arguments{
\item{deltaT}{Size of timesteps}

\item{timeSteps}{Number of timesteps we solve.  deltaT*timeSteps = total time}

\item{initialCondition}{Listing of initial conditions}

\item{variableNames}{Name of variables (i.e. P,V) that we have in our differential equation.}

\item{parameters}{The values of the parameters we are using}

\item{dynamics}{a Function that we have for our dynamics}
}
\value{
A plot of your Euler's method solution
}
\description{
\code{euler} solves a multi-dimensional differential equation with Euler's method.
See the vignette for detailed examples of usage.
}
\examples{
euler(deltaT,timeSteps,initialCondition,variableNames,FUN=dynamics,parameters=parameters)
# Run the vignette that works through an example
vignette("eulers-method")
}
