% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_temperature.R
\docType{data}
\name{global_temperature}
\alias{global_temperature}
\title{Measured average global temperature anomaly by year}
\format{
A data frame with 142 rows and 2 variables
}
\source{
The data were collected from NOAA. \url{https://climate.nasa.gov/vital-signs/global-temperature/}, download 2022-06-08
}
\usage{
data(global_temperature)
}
\description{
A dataset containing average global temperature anomaly for each year since 1880. The variables are as follows:
}
\details{
\itemize{
  \item year_since_1880. The year since 1880 (year)
  \item temperature_anomaly. Average global temperature anomaly (degrees Celsius, relative to 1951-1980)
}
}
\keyword{datasets}
