% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcEstimate.R
\name{mcmcEstimate}
\alias{mcmcEstimate}
\title{Markov Chain parameter estimates}
\usage{
mcmcEstimate(input_data, independent_var, parameters, lower_bound,
  upper_bound, iterations = 1500, burn_percent)
}
\arguments{
\item{input_data}{the data we need in order to solve our model the columns should names that correspond to the output values.}

\item{independent_var}{The name of the independent variable in our simulation}

\item{parameters}{an initial guess for our parameters}

\item{lower_bound}{the lower bound values for our parameters}

\item{upper_bound}{the upper bound values for our parameters}

\item{iterations}{the number of iterations we wish to run the MCMC for.}

\item{burn_percent}{the percentage of the iterations we discard due to "burn-in". This should be a number between 0 and 1}
}
\value{
A output of the accepted parameter histograms, model output (with uncertainty) + data utilized, and a listing of the accepted parameters.
}
\description{
\code{mcmcEstimate} Computes and Markov Chain Monte Carlo parameter estimate for a given model
}
\examples{
# Run the vignette that works through an example.
vignette("mcmc")

}
