% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phaseplane.R
\name{phaseplane}
\alias{phaseplane}
\title{Phase plane of differential equation.}
\usage{
phaseplane(
  system_eq,
  x_var,
  y_var,
  x_window = c(-4, 4),
  y_window = c(-4, 4),
  plot_points = 10,
  eq_soln = FALSE
)
}
\arguments{
\item{system_eq}{(REQUIRED) The 1 or 2 dimensional system of equations, written in formula notation as a vector (i.e.  c(dx ~ f(x,y), dy ~ g(x,y)))}

\item{x_var}{(REQUIRED) x axis variable (used in plot and formula)}

\item{y_var}{(REQUIRED) y axis variable (used in plot and formula)}

\item{x_window}{x axis limits.  Must be of the form c(minVal,maxVal).  Defaults to -4 to 4.}

\item{y_window}{y axis limits.  Must be of the form c(minVal,maxVal). Defaults to -4 to 4.}

\item{plot_points}{number of points we evaluate on the grid in both directions. Defaults to 10.}

\item{eq_soln}{TRUE / FALSE - lets you know if you want the code to estimate if there are any equilibrium solutions in the provided window}
}
\value{
A phase plane diagram of system of differential equations
}
\description{
\code{phaseplane} visualizes the vector field for a one or two dimensional differential equation.
}
\examples{
# For a two variable system of differential equations we use the formula notation for dx/dt and the dy/dt separately:
system_eq <- c(dx ~ cos(y),
              dy ~ sin(x))
phaseplane(system_eq,'x','y')

# For a one dimensional system: dy/dt = f(t,y).  In this case the xWindow represents time.
# However, the code is structured a little differently.  Consider dy/dt = -y*(1-y):

system_eq <- c(dt ~ 1,
               dy ~ -y*(1-y))

 phaseplane(system_eq,"t","y")

# Here is a second example for the differential equation dy/dt = -cos(t) y

system_eq <- c(dt ~ 1,
               dy ~ -cos(t)*y

 phaseplane(system_eq,"t","y")

In this case there may be extraneous equilibrium solutions reported due to the time dependence in the differential equations - so be sure to verify your solutions!

}
