% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phaseplane.R
\name{phaseplane}
\alias{phaseplane}
\title{Visualize a two or one dimensional phase plane for a differential equation. See the vignette for detailed examples of usage.}
\usage{
phaseplane(n_points, x_window, y_window, x_label, y_label, dx, dy)
}
\arguments{
\item{n_points}{number of points we evaluate on the grid in both directions}

\item{x_window}{x axis limits.  Must be of the form c(minVal,maxVal)}

\item{y_window}{y axis limits.  Must be of the form c(minVal,maxVal)}

\item{x_label}{x axis label for plot}

\item{y_label}{y axis label for plot}

\item{dx}{a function for the dx/dt = f(x,y)}

\item{dy}{a function for the dy/dt = f(x,y)}

\item{initialCondition}{Listing of initial conditions}
}
\value{
A phase plane diagram of linear system
}
\description{
Visualize a two or one dimensional phase plane for a differential equation. See the vignette for detailed examples of usage.
}
\examples{
dx <- function(x,y) {
cos(x)
}

dy <- function(x,y) {
 sin(y)
}

phasearrows(20,c(-3,3),c(-3,3),'X','Y',dx,dy)

# For a one dimensional system:

dx <- function(x,y) {
-x
}

phasearrows(20,c(-3,3),c(-3,3),'T','X',1,dy)

}
