% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_model.R, R/plotRegression_Data.R
\name{plotRegression_Data}
\alias{plotRegression_Data}
\title{\code{bootstrap_model} computes a given number of bootstrap samples, computes a linear model, and then plots .}
\source{
Code is adapted from: \url{https://cran.r-project.org/web/packages/broom/vignettes/bootstrapping.html}
}
\usage{
plotRegression_Data(data, regression_formula, x_label = "x",
  y_label = "y")

plotRegression_Data(data, regression_formula, x_label = "x",
  y_label = "y")
}
\arguments{
\item{data}{two column of data to be plotted.
First column is independent variable. Second column dependent variable.}

\item{regression_formula}{function you use to fit the data}

\item{x_label}{value for the x label axis}

\item{y_label}{value for the y label axis}

\item{n}{number of bootstrap samples}

\item{data}{two column of data to be plotted.
First column is independent variable. Second column dependent variable.}

\item{regression_formula}{function you use to fit the data}

\item{x_label}{value for the x label axis}

\item{y_label}{value for the y label axis}
}
\description{
\code{bootstrap_model} computes a given number of bootstrap samples, computes a linear model, and then plots .

\code{plotRegression_Data} fits a functional relationship to data, summarizes the results, and makes a plot of the fitted formula.
}
\examples{

# Identify the name of the data set you wish to run a regression on and labels
for the axis.
my_data <- data.frame(age=Loblolly$age,height=Loblolly$height)

#' Identify the basic formula of the regression and plotting formula. Make sure you have the same names of the columns in your data.
regression_formula = height ~ 1 + age+ I(age^2)

bootstrap_model(my_data,regression_formula,n=100,'Height','Age')
# Identify the name of the data set you wish to run a regression on and labels
for the axis.
my_data <- data.frame(age=Loblolly$age,height=Loblolly$height)

#' Identify the basic formula of the regression and plotting formula. Make sure you have the same names of the columns in your data.
regression_formula = height ~ 1 + age+ I(age^2)

plotRegression_Data(my_data,regression_formula,'Height','Age')
}
